requirements/dev.txt
django_filter==2.4.0
drf_yasg==1.20.0
Django==3.2
djangorestframework==3.12.4
